refactor: move misc helpers
