+++ src/sele_saisie_auto/remplir_jours_feuille_de_temps.py
@@ -41,8 +41,8 @@ from sele_saisie_auto.selenium_utils import (
     wait_for_element,
     wait_until_dom_is_stable,
 )
-from sele_saisie_auto.shared_utils import program_break_time
 from sele_saisie_auto.timeouts import DEFAULT_TIMEOUT, LONG_TIMEOUT
+from sele_saisie_auto.utils.misc import program_break_time
 
 
 @dataclass
