+++ src/sele_saisie_auto/utils/__init__.py
@@ -1,0 +1,3 @@
+"""Utility subpackages."""
+
+__all__ = ["misc"]
