+++ tests/test_saisie_automatiser_psatime.py
@@ -4,8 +4,9 @@ from pathlib import Path
 
 import pytest
 
-from sele_saisie_auto import console_ui, shared_utils
+from sele_saisie_auto import console_ui
 from sele_saisie_auto.locators import Locators
+from sele_saisie_auto.utils import misc as utils_misc
 
 sys.path.append(str(Path(__file__).resolve().parents[1] / "src"))  # noqa: E402
 
@@ -175,11 +176,11 @@ def test_helpers(monkeypatch, sample_config):
         "lundi", "8", 0, messages.TENTATIVE_INSERTION, "log.html"
     )
     monkeypatch.setattr(
-        shared_utils.subprocess,
+        utils_misc.subprocess,
         "run",
         lambda cmd, *a, **k: logs.append(cmd),
     )
-    shared_utils.clear_screen()
+    utils_misc.clear_screen()
     sap.seprateur_menu_affichage_log("log.html")
     with monkeypatch.context() as m:
         called = []
